ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
ollama==0.2.1
groq==0.4.1
python-dotenv==1.0.0
aiofiles==23.2.1
python-dateutil==2.8.2
hijri-converter==2.3.1